// src/front/SimplificaTerapia_1/SimplificaTerapia/src/app/models/paciente.ts
export interface Paciente {
    Id: number;
    nome_completo: string;
    uf: string;
    municipio: string;
    situacao: string;
    idade: number;
    cpf: string;
    rg: string;
    sexo: string;
    num_cartao_sus: string;
    telefone: string;
    responsavel: string;
    contato_responsavel: string;
    sobre: string;
    email: string;
    senha: string;
}